Add Pipeline Executor
