+++ src/sdk/python/rtdip_sdk/pipelines/_pipeline_utils/models.py
@@ -13,9 +13,13 @@
 # limitations under the License.
 
 from enum import Enum
-from typing import Optional, List
-
+from typing import Optional, List, Type, Union
 from pydantic import BaseModel
+
+from ..sources.interfaces import SourceInterface
+from ..transformers.interfaces import TransformerInterface
+from ..destinations.interfaces import DestinationInterface
+from ..utilities.interfaces import UtilitiesInterface
 
 class SystemType(Enum):
     """The type of the system."""
@@ -75,3 +79,29 @@ class Libraries(BaseModel):
                 self.add_maven_library(library)
             for library in component_libraries.pythonwheel_libraries:
                 self.add_pythonwhl_library(library)
+
+class PipelineStep(BaseModel):
+    name: str
+    description: str
+    depends_on_step: Optional[list[str]]
+    component: Union[Type[SourceInterface], Type[TransformerInterface], Type[DestinationInterface], Type[UtilitiesInterface]]
+    component_parameters: Optional[dict]
+    provide_output_to_step: Optional[list[str]]
+
+    class Config:
+        json_encoders = {
+            Union[Type[SourceInterface], Type[TransformerInterface], Type[DestinationInterface], Type[UtilitiesInterface]]: lambda x: x.__name__
+        }
+
+class PipelineTask(BaseModel):
+    name: str
+    description: str
+    depends_on_task: Optional[list[str]]
+    step_list: list[PipelineStep]
+    batch_task: Optional[bool]
+
+class PipelineJob(BaseModel):
+    name: str
+    description: str
+    version: str
+    task_list: list[PipelineTask]
